Debuggin in production.

+++ app.js
@@ -27,9 +27,10 @@ app.use((req, res, next) => {
 })
 app.use((error, req, res, next) => {
     if (req.file 
-        //&& req.route.path !=='/signup'
+        && req.route.path !=='/signup'
         ) {
         fs.unlink(req.file.path, (err) => {
+            console.log('Error validating signup.');
             console.log(err);
         })
     }
